monster commit - avro registry stuff, generic translator, mapper
add a kafka parser which reads avro schemas from the confluent registry and
decodes messages based on them. This uses the elodina/go-avro library which
produces simpler map[string]interface{} structures than the linkedin library.
The test code still contains references to the linkedin library which should be
removed in future work. It is nice to know that messages encoded by the linkedin
lib can be decoded by the elodina lib.

initial work on a generic mapper. The mapper contains three things - a Nexter, a
Translator, and a Framer (which is new). The mapper takes interface{} values and
does it's best to turn them into Pilosa records (frames, fields, row ids, column
ids) - currently it supports strings, and integers of any kind, as well as
map[string]interface{} that only contain supported values.

Nexter is now described by the INexter interface to allow for multiple
implementations (such as one based on an external entity). In the future, we'll
change the naming so that "Nexter" is the interface, and the implementation is
"LocalNexter" or something.

Moved some defintions out of source.go and proxy.go into more appropriately
named files.

Added a MapTranslator which uses an in memory map and supports arbitrary
values (except for values that cannot be map keys - byte slices excepted).