make using ioutil.ReadAll an option in taxi use case
We were using ioutil.ReadAll to ensure that we were able to download an entire
file before starting to parse and ingest it into Pilosa. This is fine if you
have a lot of memory and need perfectly repeatable results, but if you don't
have 32GB of mem and just want to try it, it's not so good. This commit makes it
an option controllable by the use-read-all flag.