remove enterprise package
This package contained support for doing imports with keys rather than ids into
the original incarnation of pilosa enterprise. We've since decided that key->id
mapping will not be an enterprise only feature, and once it is implemented, we
will bring this functionality back to PDK in a more appropriately named way.