integration test cleanup
Instead of using the "+build integration" tag which causes issues with various
automated tooling, changed the makefile target for test to use the testing.short
flag, and have integration test skip if testing.short is passed. The test-all
target does not pass the testing.short flag.

Removed some tests which were added for debugging and/or duplicated
functionality.

Added the ability to close the proxy when running the kafka integration tests