update deps, remove 'select {}' from fakeusers and default num to 1000
also changed go-pilosa dependency to be less constrained. This is causing
problems if a project using the pdk is also trying to use a particular version
of go-pilosa. I'm not actually sure that dep will be able to choose any commit
which happens to be on top of 1.1 or if it needs to be a tagged release, but
hopefully this makes things easier.